requirements.txt
python-binance==1.0.19
structlog==24.1.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
backtesting==0.3.3
fpdf2==2.7.8